修改导出个别字段取值及金额单位转换

See merge request !87

+++ middle-web/src/main/java/com/pousheng/middle/web/export/OrderExportEntity.java
@@ -76,7 +76,7 @@ public class OrderExportEntity {
      * 运费
      */
     @ExportTitle("运费")
-    private Long shipFee;
+    private Double shipFee;
 
 
     /**
@@ -116,5 +116,5 @@ public class OrderExportEntity {
      * 金额
      */
     @ExportTitle("金额")
-    private Long fee;
+    private Double fee;
 }
